added filter by user

+++ src/main/java/dev/husein/quinots/service/NoteService.java
@@ -2,19 +2,17 @@ package dev.husein.quinots.service;
 
 import com.google.common.base.Strings;
 import dev.husein.quinots.adapter.TagsConverter;
-import dev.husein.quinots.controller.ExceptionsController;
+import dev.husein.quinots.dto.NoteDTO;
 import dev.husein.quinots.exception.QuinotsException;
 import dev.husein.quinots.model.BaseResponse;
 import dev.husein.quinots.model.Note;
 import dev.husein.quinots.model.User;
 import dev.husein.quinots.repository.NoteRepository;
-import dev.husein.quinots.repository.UserRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.sql.Timestamp;
 import java.util.*;
 
 @Service
@@ -30,24 +28,36 @@ public class NoteService {
         this.userService = userService;
     }
 
-    public BaseResponse createNote(Long userId, String title, String description, Timestamp dateTimeCreated, List<String> tags) {
+    public BaseResponse createNote(String username, String title, String description, List<String> tags) {
         LOGGER.debug("Creating Note");
-        LOGGER.debug("\nTitle: {} \nDescription: {} \nDate-Time Created: {} \nTags: {}", title, description, dateTimeCreated, tags);
-        LOGGER.debug("User: {}", userId);
+        LOGGER.debug("\nTitle: {} \nDescription: {} \nTags: {}", title, description, tags);
+        LOGGER.debug("User: {}", username);
 
         BaseResponse baseResponse = new BaseResponse();
-        baseResponse.addNoteToList(noteRepository.saveAndFlush(new Note(title, description, dateTimeCreated, tags, verifyUser(userId))));
+        baseResponse.addNoteToList(noteRepository.saveAndFlush(new Note(title, description, tags, userService.verifyUser(username))));
         return baseResponse;
     }
 
-    public void deleteNote(Long id) {
-        noteRepository.deleteById(id);
+    public void deleteNote(String username, Long id) {
+        Long userId = userService.getUserIdByUsername(username);
+        Note note = noteRepository.getById(id);
+        if (userId.equals(note.getUser().getId())) {
+            noteRepository.deleteById(id);
+        } else {
+            throw new QuinotsException(String.format("Note with ID %d can't be found", id));
+        }
     }
 
-    public void updateNote(Long id, Note newEditedNote) {
+    public void updateNote(String username, Long id, NoteDTO newEditedNote) {
+        Long userId = userService.getUserIdByUsername(username);
+
         if (noteRepository.existsById(id)) {
             Note existingNote = noteRepository.getById(id);
             // check every field to make sure you don't override the field if it is empty
+
+            if (!userId.equals(existingNote.getUser().getId())) {
+                throw new QuinotsException(String.format("Note with ID %d can't be found", id));
+            }
 
             if (!Strings.isNullOrEmpty(newEditedNote.getTitle())) {
                 existingNote.setTitle(newEditedNote.getTitle());
@@ -64,15 +74,15 @@ public class NoteService {
             noteRepository.saveAndFlush(existingNote);
 
         } else {
-            throw new QuinotsException("Error updating note");
+            throw new QuinotsException(String.format("Note with ID %d can't be found", id));
         }
     }
 
-    public BaseResponse listAllNotes() {
-        return new BaseResponse(noteRepository.findAll());
+    public BaseResponse listNotesForUser(String username) {
+        return new BaseResponse(noteRepository.listNotesForUser(userService.getUserIdByUsername(username)));
     }
 
-    public BaseResponse searchNotes(String includingWords, Calendar fromDate, Calendar toDate, String tagsStr) {
+    public BaseResponse searchNotes(String username, String includingWords, Calendar fromDate, Calendar toDate, String tagsStr) {
         BaseResponse baseResponse = new BaseResponse();
         List<String> tags = new ArrayList<>();
 
@@ -80,7 +90,7 @@ public class NoteService {
             tags = Arrays.asList(tagsStr.split(TagsConverter.SEPARATOR_CHAR));
         }
 
-        List<Note> listOfNotesWithAnyTag = noteRepository.searchNotes(includingWords, fromDate, toDate);
+        List<Note> listOfNotesWithAnyTag = noteRepository.searchNotes(userService.getUserIdByUsername(username), includingWords, fromDate, toDate);
 
         // if tag filter is requested
         if (tags.size() > 0) {
@@ -104,21 +114,20 @@ public class NoteService {
         return baseResponse;
     }
 
-    public BaseResponse getNoteById(Long id) {
+    public BaseResponse getNoteById(String username, Long id) {
+        Long userId = userService.getUserIdByUsername(username);
         BaseResponse baseResponse = new BaseResponse();
         if (noteRepository.existsById(id)) {
-            baseResponse.addNoteToList(noteRepository.findById(id).orElse(null));
+            Note note = noteRepository.findById(id).orElse(null);
+            if (userId.equals(note.getUser().getId()))
+                baseResponse.addNoteToList(note);
+            else
+                throw new QuinotsException(String.format("Note with ID %d can't be found", id));
         } else {
             throw new QuinotsException(String.format("Note with ID %d can't be found", id));
         }
 
         return baseResponse;
-    }
-
-    private User verifyUser(Long userId) throws NoSuchElementException {
-        return userService.findUserById(userId).orElseThrow(() ->
-                new NoSuchElementException("User does not exist " + userId)
-        );
     }
 
 }
